Use Separate Account for Prefunding Trader
This PR builds on #1 and uses a separate account for pre-funding the
trader with sell tokens in case they don't have a sufficient balance.
This was done to interfere less with the settlement process:

> Also technically we should increase the solver sell token balance by
> the needed amount. Technically a solver could have private inventory
> they would like to use for the solution instead of transfering it to
> the trader.

Because of this, it is not safe to override or use the `solver`'s sell
token balance (it might be used during the settlement). This does mean
that we need "simulation magic" for this feature to work in a way that
does not intefere with the settlement.