Add A Balance Override Detector
This PR adds a balance override detector for quote verification. This
allows

The current implementation strategy is quite simple - just try a bunch
of different slots and see which one works. For now, the number of slots
to try is hardcoded at 25, but this can be configurable in the future.

The component is marked with `#[allow(dead_code)]` as it isn't being
used anywhere, In the interest of keeping the PR size down, I will add
all of the glue code to make it used in a follow up (which will also
test whether or not the detector works as expected).

In the meantime, I did some local testing to check that it works with
tokens for which I know the mapping slot on Ethereum Mainnet (test not
included in the PR).